Explicit converion to float32 of the logits argument of tf.nn.ctc_loss

+++ model/keras_model.py
@@ -82,7 +82,7 @@ def CTCLoss(labels, logits, features_length, input_length, labels_length):
     # https://www.tensorflow.org/api_docs/python/tf/nn/ctc_loss
     return tf.reduce_mean(tf.nn.ctc_loss(
         tf.cast(labels, dtype=tf.int32),
-        tf.cast(logits, dtype=tf.int32), 
+        tf.cast(logits, dtype=tf.float32), 
         tf.cast(labels_length, dtype=tf.int32), 
         ctc_input_length, 
         logits_time_major=False))
